:recycle: move parsing to utils/
