+++ router/events.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/traPtitech/knoQ/domain"
 	"github.com/traPtitech/knoQ/domain/filter"
-	"github.com/traPtitech/knoQ/parsing"
 	"github.com/traPtitech/knoQ/presentation"
+	"github.com/traPtitech/knoQ/utils/parsing"
 
 	"github.com/gofrs/uuid"
